Find And Print The sum Of All Even FibonacciTerms Below 4000000

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,11 +1,18 @@
 #include <stdio.h>
+#include "main.h"
+/**
+ *
+ * Description: Prints The even fibonacci seequence below 4000000,
+ * Followed by a new line.
+ * Return: Always 0.
+ */
 
 int main(void)
 {
-    long long a = 0; // initialize a as the first Fibonacci number
-    long long b = 1; // initialize b as the second Fibonacci number
-    long long c = 0; // initialize c as the next Fibonacci number
-    long long sum = 0; // initialize sum to keep track of the even Fibonacci numbers
+    long a = 0; // initialize a as the first Fibonacci number
+    long b = 1; // initialize b as the second Fibonacci number
+    long c = 0; // initialize c as the next Fibonacci number
+    long sum = 0; // initialize sum to keep track of the even Fibonacci numbers
 
     while (c <= 4000000) { // continue looping as long as c is less than 4,000,000
         c = a + b; // calculate the next Fibonacci number
@@ -15,7 +22,7 @@ int main(void)
             sum += c; // add c to the sum
         }
     }
-    printf("%lld\n", sum); // print the sum of the even Fibonacci numbers
+    printf("%ld\n", sum); // print the sum of the even Fibonacci numbers
     return (0);
 }
